+++ tests/integration/TransactionTest.php
@@ -2809,6 +2809,55 @@ class TransactionTest extends Setup
         $this->assertEquals('67.00', $submitResult->transaction->amount);
     }
 
+  public function testSubmitForSettlement_withLevel2Data()
+    {
+        $transaction = Braintree\Transaction::saleNoValidate([
+            'amount' => '100.00',
+            'creditCard' => [
+                'number' => '5105105105105100',
+                'expirationDate' => '05/12'
+            ]
+        ]);
+
+        $this->assertEquals(Braintree\Transaction::AUTHORIZED, $transaction->status);
+        $submitForSettlementParams = [
+            'purchaseOrderNumber' => 'ABC123',
+            'taxAmount' => '1.34',
+            'taxExempt' => true
+        ];
+        $submitResult = Braintree\Transaction::submitForSettlement($transaction->id, null, $submitForSettlementParams);
+        $this->assertEquals(true, $submitResult->success);
+        $this->assertEquals(Braintree\Transaction::SUBMITTED_FOR_SETTLEMENT, $submitResult->transaction->status);
+    }
+
+  public function testSubmitForSettlement_withLevel3Data()
+    {
+        $transaction = Braintree\Transaction::saleNoValidate([
+            'amount' => '100.00',
+            'creditCard' => [
+                'number' => '5105105105105100',
+                'expirationDate' => '05/12'
+            ]
+        ]);
+
+        $this->assertEquals(Braintree\Transaction::AUTHORIZED, $transaction->status);
+        $submitForSettlementParams = [
+          'shippingAmount' => '1.00',
+          'discountAmount' => '2.00',
+          'shipsFromPostalCode' => '12345',
+          'lineItems' => [[
+              'quantity' => '1.0232',
+              'name' => 'Name #1',
+              'kind' => Braintree\TransactionLineItem::DEBIT,
+              'unitAmount' => '45.1232',
+              'totalAmount' => '45.15',
+          ]]
+        ];
+        $submitResult = Braintree\Transaction::submitForSettlement($transaction->id, null, $submitForSettlementParams);
+        $this->assertEquals(true, $submitResult->success);
+        $this->assertEquals(Braintree\Transaction::SUBMITTED_FOR_SETTLEMENT, $submitResult->transaction->status);
+    }
+
   public function testSubmitForSettlement_withDescriptor()
     {
         $transaction = Braintree\Transaction::saleNoValidate([
@@ -4054,6 +4103,30 @@ class TransactionTest extends Setup
         );
     }
 
+  public function testHandlesSoftDeclinedRefundAuth()
+    {
+        $transaction = $this->createTransactionToRefundAuth();
+        $result = Braintree\Transaction::refund($transaction->id, '2046.00');
+        $this->assertFalse($result->success);
+        $errors = $result->errors->forKey('transaction')->onAttribute('base');
+        $this->assertEquals(
+            Braintree\Error\Codes::TRANSACTION_REFUND_AUTH_SOFT_DECLINED,
+            $errors[0]->code
+        );
+    }
+
+  public function testHandlesHardDeclinedRefundAuth()
+    {
+        $transaction = $this->createTransactionToRefundAuth();
+        $result = Braintree\Transaction::refund($transaction->id, '2009.00');
+        $this->assertFalse($result->success);
+        $errors = $result->errors->forKey('transaction')->onAttribute('base');
+        $this->assertEquals(
+            Braintree\Error\Codes::TRANSACTION_REFUND_AUTH_HARD_DECLINED,
+            $errors[0]->code
+        );
+    }
+
   public function testRefundWithOptionsParam()
     {
         $transaction = $this->createTransactionToRefund();
@@ -4305,6 +4378,17 @@ class TransactionTest extends Setup
                 'number' => '5105105105105100',
                 'expirationDate' => '05/12'
             ],
+            'options' => ['submitForSettlement' => true]
+        ]);
+        Braintree\Test\Transaction::settle($transaction->id);
+        return $transaction;
+    }
+
+  public function createTransactionToRefundAuth()
+    {
+        $transaction = Braintree\Transaction::saleNoValidate([
+            'amount' => '9000.00',
+            'paymentMethodNonce' => Braintree\Test\Nonces::$transactable,
             'options' => ['submitForSettlement' => true]
         ]);
         Braintree\Test\Transaction::settle($transaction->id);